# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NSG MOA API Documentation",
        "description": "Nasarawa State Ministry of Agriculture M&E System API documentation",
        "contact": {
            "email": "support@nsgmoa.gov.ng"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/monitoring/queue": {
            "get": {
                "tags": [
                    "Monitoring"
                ],
                "summary": "Get queue status",
                "operationId": "97c5b1099abf2beb5b35ac64214fdee8",
                "responses": {
                    "200": {
                        "description": "Queue status information"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/monitoring/sms": {
            "get": {
                "tags": [
                    "Monitoring"
                ],
                "summary": "Get SMS statistics",
                "operationId": "e502446495c676933abb2709d44465ff",
                "responses": {
                    "200": {
                        "description": "SMS statistics"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/monitoring/failed-jobs": {
            "get": {
                "tags": [
                    "Monitoring"
                ],
                "summary": "Get failed jobs",
                "operationId": "12fc728d4bcbd4647bac08cf1e485cfd",
                "responses": {
                    "200": {
                        "description": "Failed jobs list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/monitoring/retry-job/{id}": {
            "post": {
                "tags": [
                    "Monitoring"
                ],
                "summary": "Retry a failed job",
                "operationId": "b40e0ebdcf84f40d1c3ed968e292fad5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job retry status"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}